Rename old actions tests to deprecated_actions
