+++ lib/dal/tests/integration_test/mod.rs
@@ -1,4 +1,3 @@
-mod action;
 mod attribute;
 mod before_funcs;
 mod builtins;
@@ -7,6 +6,7 @@ mod component;
 mod connection;
 mod cycle_check_guard;
 mod dependent_values_update;
+mod deprecated_action;
 mod frame;
 mod func;
 mod input_sources;
